requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
Django==4.0.9
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
inflection==0.5.1
pycodestyle==2.9.1
pytz==2022.2.1
sqlparse==0.4.2
toml==0.10.2
gunicorn==20.1.0
whitenoise==6.1.0
python-dotenv==0.21.0
